Fix RequestFactory gsub for base_path_from_servers
Commit 8fdd3bf979b10b4c65bc6ce48bc2a7b5eb02651b introduced a change to this method
where the gsub replacement of the server url switched from the global Regexp match
variable of \$1 to the block argument for gsub.
This change made it so that the argument to the variable name lookup include the
braces used in the string for the url variable.
That lead to the hash not returing any matching keys and values and the generated
URI string of '://' being passed to the URI() method which raise an error.